+++ MatrixSDK/Crypto/Algorithms/Megolm/MXMegolmEncryption.m
@@ -26,47 +26,7 @@
 #import "MXCrypto_Private.h"
 #import "MXQueuedEncryption.h"
 #import "MXTools.h"
-
-@interface MXOutboundSessionInfo : NSObject
-{
-    // When the session was created
-    NSDate  *creationTime;
-}
-
-- (instancetype)initWithSessionID:(NSString*)sessionId;
-
-/**
- Check if it's time to rotate the session.
-
- @param rotationPeriodMsgs the max number of encryptions before rotating.
- @param rotationPeriodMs the max duration of an encryption session before rotating.
- @return YES if rotation is needed.
- */
-- (BOOL)needsRotation:(NSUInteger)rotationPeriodMsgs rotationPeriodMs:(NSUInteger)rotationPeriodMs;
-
-/**
- Determine if this session has been shared with devices which it shouldn't
- have been.
-
- @param devicesInRoom userId -> {deviceId -> object} devices we should shared the session with.
- @return YES if we have shared the session with devices which aren't in devicesInRoom.
- */
-- (BOOL)sharedWithTooManyDevices:(MXUsersDevicesMap<MXDeviceInfo *> *)devicesInRoom;
-
-// The id of the session
-@property (nonatomic, readonly) NSString *sessionId;
-
-// Number of times this session has been used
-@property (nonatomic) NSUInteger useCount;
-
-// If a share operation is in progress, the corresping http request
-@property (nonatomic) MXHTTPOperation* shareOperation;
-
-// Devices with which we have shared the session key
-// userId -> {deviceId -> msgindex}
-@property (nonatomic) MXUsersDevicesMap<NSNumber*> *sharedWithDevices;
-
-@end
+#import "MXOutboundSessionInfo.h"
 
 
 @interface MXMegolmEncryption ()
@@ -123,6 +83,13 @@
         // TODO: Make it configurable via parameters
         sessionRotationPeriodMsgs = 100;
         sessionRotationPeriodMs = 7 * 24 * 3600 * 1000;
+        
+        // restore last saved outbound session for this room
+        outboundSession = [crypto.olmDevice restoreOutboundGroupSessionForRoom:roomId];
+        if (outboundSession)
+        {
+            outboundSessions[outboundSession.sessionId] = outboundSession;
+        }
     }
     return self;
 }
@@ -278,12 +245,18 @@
     // Need to make a brand new session?
     if (session && [session needsRotation:sessionRotationPeriodMsgs rotationPeriodMs:sessionRotationPeriodMs])
     {
+        [crypto.olmDevice discardOutboundGroupSessionForRoom:roomId];
+        [outboundSessions removeObjectForKey:session.sessionId];
+        outboundSession = nil;
         session = nil;
     }
 
     // Determine if we have shared with anyone we shouldn't have
     if (session && [session sharedWithTooManyDevices:devicesInRoom])
     {
+        [crypto.olmDevice discardOutboundGroupSessionForRoom:roomId];
+        [outboundSessions removeObjectForKey:session.sessionId];
+        outboundSession = nil;
         session = nil;
     }
 
@@ -336,7 +309,7 @@
 
 - (MXOutboundSessionInfo*)prepareNewSession
 {
-    NSString *sessionId = [crypto.olmDevice createOutboundGroupSession];
+    NSString *sessionId = [crypto.olmDevice createOutboundGroupSessionForRoomWithId:self->roomId];
 
     [crypto.olmDevice addInboundGroupSession:sessionId
                                   sessionKey:[crypto.olmDevice sessionKeyForOutboundGroupSession:sessionId]
@@ -568,6 +541,9 @@
                       });
 
             session.useCount++;
+            
+            // We have to store the session in the DB est time a message is encrypted to save the session useCount
+            [crypto.olmDevice storeOutboundGroupSessionForRoomWithId:roomId withSessionId:session.sessionId];
         }
     }
     else
@@ -579,62 +555,6 @@
     }
 
     [pendingEncryptions removeAllObjects];
-}
-
-@end
-
-
-#pragma mark - MXOutboundSessionInfo
-
-@implementation MXOutboundSessionInfo
-
-- (instancetype)initWithSessionID:(NSString *)sessionId
-{
-    self = [super init];
-    if (self)
-    {
-        _sessionId = sessionId;
-        _sharedWithDevices = [[MXUsersDevicesMap alloc] init];
-        creationTime = [NSDate date];
-    }
-    return self;
-}
-
-- (BOOL)needsRotation:(NSUInteger)rotationPeriodMsgs rotationPeriodMs:(NSUInteger)rotationPeriodMs
-{
-    BOOL needsRotation = NO;
-    NSUInteger sessionLifetime = [[NSDate date] timeIntervalSinceDate:creationTime] * 1000;
-
-    if (_useCount >= rotationPeriodMsgs || sessionLifetime >= rotationPeriodMs)
-    {
-        NSLog(@"[MXMegolmEncryption] Rotating megolm session after %tu messages, %tu ms", _useCount, sessionLifetime);
-        needsRotation = YES;
-    }
-
-    return needsRotation;
-}
-
-- (BOOL)sharedWithTooManyDevices:(MXUsersDevicesMap<MXDeviceInfo *> *)devicesInRoom
-{
-    for (NSString *userId in _sharedWithDevices.userIds)
-    {
-        if (![devicesInRoom deviceIdsForUser:userId])
-        {
-            NSLog(@"[MXMegolmEncryption] Starting new session because we shared with %@",  userId);
-            return YES;
-        }
-
-        for (NSString *deviceId in [_sharedWithDevices deviceIdsForUser:userId])
-        {
-            if (! [devicesInRoom objectForDevice:deviceId forUser:userId])
-            {
-                NSLog(@"[MXMegolmEncryption] Starting new session because we shared with %@:%@", userId, deviceId);
-                return YES;
-            }
-        }
-    }
-
-    return NO;
 }
 
 @end